+++ martin/src/pg/pg_source.rs
@@ -1,5 +1,3 @@
-use std::collections::HashMap;
-
 use async_trait::async_trait;
 use deadpool_postgres::tokio_postgres::types::{ToSql, Type};
 use log::debug;
@@ -58,35 +56,32 @@ impl Source for PgSource {
 
     async fn get_tile(
         &self,
-        xyz: &TileCoord,
-        url_query: &Option<UrlQuery>,
+        xyz: TileCoord,
+        url_query: Option<&UrlQuery>,
     ) -> MartinResult<TileData> {
-        let empty_query = HashMap::new();
-        let url_query = url_query.as_ref().unwrap_or(&empty_query);
         let conn = self.pool.get().await?;
-
         let param_types: &[Type] = if self.support_url_query() {
             &[Type::INT2, Type::INT8, Type::INT8, Type::JSON]
         } else {
             &[Type::INT2, Type::INT8, Type::INT8]
         };
 
-        let query = &self.info.query;
+        let sql = &self.info.sql_query;
         let prep_query = conn
-            .prepare_typed_cached(query, param_types)
+            .prepare_typed_cached(sql, param_types)
             .await
             .map_err(|e| {
                 PrepareQueryError(
                     e,
                     self.id.to_string(),
                     self.info.signature.to_string(),
-                    self.info.query.to_string(),
+                    self.info.sql_query.to_string(),
                 )
             })?;
 
         let tile = if self.support_url_query() {
             let json = query_to_json(url_query);
-            debug!("SQL: {query} [{xyz}, {json:?}]");
+            debug!("SQL: {sql} [{xyz}, {json:?}]");
             let params: &[&(dyn ToSql + Sync)] = &[
                 &i16::from(xyz.z),
                 &i64::from(xyz.x),
@@ -95,7 +90,7 @@ impl Source for PgSource {
             ];
             conn.query_opt(&prep_query, params).await
         } else {
-            debug!("SQL: {query} [{xyz}]");
+            debug!("SQL: {sql} [{xyz}]");
             conn.query_opt(
                 &prep_query,
                 &[&i16::from(xyz.z), &i64::from(xyz.x), &i64::from(xyz.y)],
@@ -107,9 +102,9 @@ impl Source for PgSource {
             .map(|row| row.and_then(|r| r.get::<_, Option<TileData>>(0)))
             .map_err(|e| {
                 if self.support_url_query() {
-                    GetTileWithQueryError(e, self.id.to_string(), *xyz, url_query.clone())
+                    GetTileWithQueryError(e, self.id.to_string(), xyz, url_query.cloned())
                 } else {
-                    GetTileError(e, self.id.to_string(), *xyz)
+                    GetTileError(e, self.id.to_string(), xyz)
                 }
             })?
             .unwrap_or_default();
@@ -120,7 +115,7 @@ impl Source for PgSource {
 
 #[derive(Clone, Debug)]
 pub struct PgSqlInfo {
-    pub query: String,
+    pub sql_query: String,
     pub use_url_query: bool,
     pub signature: String,
 }
@@ -129,7 +124,7 @@ impl PgSqlInfo {
     #[must_use]
     pub fn new(query: String, has_query_params: bool, signature: String) -> Self {
         Self {
-            query,
+            sql_query: query,
             use_url_query: has_query_params,
             signature,
         }
