+++ martin/src/srv/mod.rs
@@ -2,7 +2,16 @@ mod config;
 pub use config::{SrvConfig, KEEP_ALIVE_DEFAULT, LISTEN_ADDRESSES_DEFAULT};
 
 mod server;
-pub use server::{
-    get_tile_content, get_tile_response, merge_tilejson, new_server, router, Catalog, TileRequest,
-    RESERVED_KEYWORDS,
-};
+pub use server::{new_server, router, Catalog, RESERVED_KEYWORDS};
+
+mod tiles;
+pub use tiles::{get_tile_content, get_tile_response, TileRequest};
+
+#[cfg(feature = "fonts")]
+mod fonts;
+
+mod tiles_info;
+pub use tiles_info::{merge_tilejson, SourceIDsRequest};
+
+#[cfg(feature = "sprites")]
+mod sprites;
