+++ martin/tests/pg_function_source_test.rs
@@ -32,14 +32,14 @@ async fn function_source_tile() {
     let mock = mock_sources(mock_pgcfg("connection_string: $DATABASE_URL")).await;
     let src = source(&mock, "function_zxy_query");
     let tile = src
-        .get_tile(&TileCoord { z: 0, x: 0, y: 0 }, &None)
+        .get_tile(TileCoord { z: 0, x: 0, y: 0 }, None)
         .await
         .unwrap();
     assert!(!tile.is_empty());
 
     let src = source(&mock, "function_zxy_query_jsonb");
     let tile = src
-        .get_tile(&TileCoord { z: 0, x: 0, y: 0 }, &None)
+        .get_tile(TileCoord { z: 0, x: 0, y: 0 }, None)
         .await
         .unwrap();
     assert!(!tile.is_empty());
